requirements.txt
aiosqlite==0.17.0
aniso8601==9.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
asyncpg==0.27.0
attrs==22.1.0
autopep8==2.0.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
CacheControl==0.12.11
cachetools==5.2.0
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ciso8601==2.2.0
cleo==1.0.0a5
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
crashtest==0.3.1
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dnspython==2.2.1
docutils==0.19
dulwich==0.20.50
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
executing==1.2.0
fastapi==0.87.0
fastapi-jwt==0.1.8
fastjsonschema==2.16.2
filelock==3.8.0
firebase==3.0.1
firebase-admin==6.0.1
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.38.0
gcloud==0.18.3
geolocation-python==0.2.2
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
googlemaps==4.7.0
gps3==0.33.3
greenlet==2.0.1
grpcio==1.50.0
grpcio-status==1.50.0
h11==0.14.0
html5lib==1.1
httpcore==0.16.1
httplib2==0.21.0
httpx==0.23.1
idna==3.4
importlib-metadata==4.13.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
iso8601==1.1.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.1
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
jws==0.1.3
jwt==1.3.1
keyring==23.11.0
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.4
lockfile==0.12.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==9.0.0
msgpack==1.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauth2client==4.1.3
orjson==3.8.1
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkginfo==1.8.3
platformdirs==2.5.4
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pylev==1.4.0
pymongo==4.3.3
pyparsing==3.0.9
pypika-tortoise==0.1.6
Pyrebase4==4.6.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-firebase==1.2
python-jose==3.3.0
python-jwt==2.0.1
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
seaborn==0.12.1
SecretStorage==3.3.3
Send2Trash==1.8.0
shellingham==1.5.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlmodel==0.0.8
sseclient==0.0.27
stack-data==0.6.1
starlette==0.21.0
terminado==0.17.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tortoise-orm==0.19.2
traitlets==5.5.0
typer==0.7.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
virtualenv==20.16.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
zipp==3.10.0
